+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - cmc `v0.0.4`
 - utils `v0.0.7`
 
+### Features
+
+- new Sns features: increase stake neuron
+
 ### Build
 
 - bump `agent-js` peer dependencies to `v0.15.1`